Added contributing.md and re-organized packages to prepare for future re-factoring
