--- .github/workflows/benchmark.yml
@@ -1,21 +1,0 @@
-name: benchmark
-
-on: [ pull_request ]
-
-jobs:
-  benchmark:
-    name: Run benchmarks
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Set up Go
-        uses: actions/setup-go@v2
-        with:
-          go-version: 1.16
-
-      - name: Install dependencies
-        run: go get ./...
-
-      - name: Run benchmarks
-        run: go test -bench . -benchmem ./...
