+++ cmd/pillager/root.go
@@ -1,4 +1,4 @@
-package cmd
+package pillager
 
 import (
 	"fmt"
@@ -32,7 +32,7 @@ var rootCmd = &cobra.Command{
 }
 
 // Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
+// This is called by pillager.pillager(). It only needs to happen once to the rootCmd.
 func Execute() {
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Println(err)
