+++ pkg/hunter/docs.go
