+++ pkg/hunter/template.go
