+++ pkg/rules/README.md
